requirements/dev.txt
flake8==6.1.0
flake8-import-order==0.18.2
flake8-quotes==3.3.2
pep8-naming==0.13.3
black==23.9.1
django-debug-toolbar==4.2.0